Add first implementation of sparse kernels.
Note: sparse kernel sensesum cannot be used when more than one entry connects the same pair,
however this is caught as an error.

Fixed bug with compare_against='triton'.
