+++ hippynn/custom_kernels/test_env_cupy.py
@@ -1,8 +1,7 @@
-import torch
-import numba
+
 from . import env_cupy
-from .test_env_numba import Envops_tester, main
-from .test_env_numba import TEST_MEGA_PARAMS, TEST_LARGE_PARAMS, TEST_MEDIUM_PARAMS, TEST_SMALL_PARAMS
+from .test_env import main
+
 
 if __name__ == "__main__":
     main(
