+++ hippynn/custom_kernels/test_env_triton.py
@@ -1,13 +1,11 @@
-import torch
-from .env_triton import envsum, sensesum, featsum
-from .test_env_numba import Envops_tester, main, get_simulated_data
-from .test_env_numba import TEST_MEGA_PARAMS, TEST_LARGE_PARAMS, TEST_MEDIUM_PARAMS, TEST_SMALL_PARAMS
-from .utils import resort_pairs_cached
+
+from . import env_triton
+from .test_env import main
 
 if __name__ == "__main__":
 
     main(
-        envsum,
-        sensesum,
-        featsum,
+        env_triton.envsum,
+        env_triton.sensesum,
+        env_triton.featsum,
     )
